move metoken util out of setup folder; move oracle tests to oracle file
